Added the actual project folders
Somehow, none of us realized that we did not have the actual project solution in the repository until the entire project was completed. This is the repository without any of the .h or .cpp files. I will now attempt to move the .h and .cpp files from the repository into the correct folder.